🎁Modify: RoomCard, DormReviewCard 컴포넌트 분리
